bulkmailzip/bulk_mail/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
bcrypt==4.1.3
beautifulsoup4==4.13.4
blinker==1.8.2
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.4.1
greenlet==3.0.3
gunicorn
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
mysql-connector-python==8.4.0
openpyxl==3.1.5
packaging==24.1
pdfkit==1.0.0
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.2
six==1.16.0
soupsieve==2.7
SQLAlchemy==2.0.30
test-rest-api==0.0.0.0.40
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
WTForms==3.1.2
yarl==1.9.4
